Sépare le dépot de données en plusieurs sous-dépôts
